// src/webparts/SharePointThemeManager/models/Palette.ts
export interface Palette {
    themeLighter: string;
    themeSecondary: string;
    themeDarkAlt: string;
    neutralTertiaryAlt: string;
    themePrimary: string;
    themeTertiary: string;
    themeDarker: string;
    neutralLighter: string;
    neutralTertiary: string;
    neutralLight: string;
    themeLighterAlt: string;
    neutralLighterAlt: string;
    black: string;
    neutralPrimaryAlt: string;
    neutralQuaternaryAlt: string;
    themeDark: string;
    white: string;
    neutralDark: string;
    neutralSecondary: string;
    neutralPrimary: string;
    themeLight: string;
    neutralQuaternary: string;
}